refactor: move nickname validation to dedicated module
- Refactored the `validate_nickname` function from `validation_utils.py` to a new module `nickname_validation.py` for better organization and clarity.
- Updated import statements in `user_api.py` and the corresponding test file to reflect the new location of the `validate_nickname` function.

This change enhances code maintainability by separating concerns related to nickname validation.
